pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Authentication micro-service for user from telegram using JWT tokens"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "black>=25.1.0",
    "coverage>=7.9.2",
    "django>=5.2.4",
    "django-rest-framework>=0.1.0",
    "djangorestframework-simplejwt>=5.5.0",
    "dotenv>=0.9.9",
    "psycopg2>=2.9.10",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-django>=4.11.1",
    "ruff>=0.12.4",
]
